Command-line version now uses twisted in order to improve range requests.
I don't particularly like twisted.  I think gevent is a much better
server, but twisted is already a dependency to build the desktop version,
and I don't feel like doing much work right now to use gevent.

Also added support for the "--back" command-line argument which allows
the caller to push to the back of the queue.  Without this argument,
flingo.py pushes to the front of the user's queue.